refactor: create context folder move context from signup
