+++ src/routes/sign-up/steps/connect-wallet-scan/body.jsx
@@ -1,7 +1,8 @@
 //connect-wallet-scan-body
 import { useContext, useEffect, useState } from "react";
+import { WizardContext } from "../../../../contexts/wizard-context";
 import { account } from "../../../../utils/faker-data";
-import { WizardContext } from "../../sign-up";
+// import { WizardContext } from "../../sign-up";
 
 //TODO: What happens with the camera component
 //QR Code will have information
@@ -16,6 +17,7 @@ import { WizardContext } from "../../sign-up";
  * @param WizardContext context for data in Sign up Wizard
  * @param updateAccountData function to update scan user data in Sign up Wizard Context
  * @param seedPhrases seed phrases from scan to update in Signup Wizard Context
+ * @param updateSeedPhrases function to update SignupWizardContext
  * @param data useState value containing data from scan
  * @param setData useState setter to set data from scan
  * @param account eximported faker user data to emulate scan
@@ -27,7 +29,8 @@ import { WizardContext } from "../../sign-up";
 export const Body = () => {
   /* ----------------------------- set up Context ----------------------------- */
   //get update scanAccountData
-  const { updateAccountData, seedPhrases } = useContext(WizardContext);
+  const { updateAccountData, seedPhrases, updateSeedPhrases } =
+    useContext(WizardContext);
   /* -------------------------------- useState -------------------------------- */
   const [data, setData] = useState({});
